Move engines to submodule, enable based on metadata
